+++ POO/src/ComponentesSistema/Veiculos/CarroHibrido.java
@@ -10,7 +10,7 @@ public class CarroHibrido extends VeiculoComAutonomia {
     public CarroHibrido (String nova_matricula,
                           double velMedia,
                           double priceKM,
-                          double classiF,
+                          int classiF,
                           Localizacao nova_local,
                           double consumo,
                           double maxAuto,
