+++ POO/src/ComponentesSistema/Veiculos/VeiculoComAutonomia.java
@@ -19,7 +19,7 @@ public class VeiculoComAutonomia extends Veiculo {
     							String nova_matricula,
 		                        double velMedia,
 		                        double priceKM,
-		                        double classiF,
+		                        int classiF,
 		                        Localizacao nova_local,
 		                        double consumo,
 		                        double maxAuto,
